Command history and color

+++ src/util/camera.cpp
@@ -29,8 +29,8 @@ void Camera::update(float dt)
 	if (Input::keys[GLFW_KEY_LEFT_CONTROL]) mv -= -up * speed *  dt;
 	view = glm::translate(glm::mat4(), mv) * view;
 
-	float dx = Input::mouseX - last_x;
-	float dy = Input::mouseY - last_y;
+	double dx = Input::mouseX - last_x;
+	double dy = Input::mouseY - last_y;
 	view = glm::rotate((float) dy * dt, glm::vec3(1.0f, 0.0f, 0.0f)) * view;
 	view = glm::rotate((float) dx * dt, glm::vec3(0.0f, 1.0f, 0.0f)) * view;
 	if (Input::keys[GLFW_KEY_Q]) view = glm::rotate(glm::radians(-5.0f) * speed *  dt, glm::vec3(0.0f, 0.0f, 1.0f)) * view;
